# obsidian-vault/wiki/summaries/API契约/OpenAPI-接口契约-v2.2.yaml
openapi: 3.0.3
info:
  title: 图文快印门店连锁管理系统 Print MIS API
  version: 2.2.0
  description: 准生产版 v2 接口契约，在 v2.1 变更冻结基础上追加统一工作流动作、角色任务工作台和订单聚合详情。
servers:
  - url: http://127.0.0.1:8080
security:
  - basicAuth: []
...
paths:
...
  /api/orders/{orderId}/workflow/actions/{action}:
    post:
      tags: [Order]
      summary: 统一订单工作流动作入口
      description: 统一承接提交审核、报价、作业单、排产、完工、配送、签收、收款、开票、退款和申请变更等动作；旧快捷接口保留兼容。
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [SUBMIT_REVIEW, QUOTE, JOB_TICKET, SCHEDULE_PRODUCTION, COMPLETE_PRODUCTION, CREATE_DELIVERY, ACCEPT_DELIVERY, SIGN_DELIVERY, PAY, INVOICE, REFUND, REQUEST_CHANGE]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 返回订单、动作结果、下一步任务和消息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/production-tasks:
    get:
      tags: [Production]
      summary: 生产任务列表
      responses:
        '200':
          description: 生产任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Production]
      summary: 创建生产任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionTask'
      responses:
        '200':
          description: 创建后的生产任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/production-tasks/{id}:
    get:
      tags: [Production]
      summary: 生产任务详情
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 生产任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Production]
      summary: 更新生产任务
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionTask'
      responses:
        '200':
          description: 更新后的生产任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Production]
      summary: 删除生产任务
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已删除生产任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/production-tasks/{id}/complete:
    post:
      tags: [Production]
      summary: 生产完工并质检通过
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 更新后的生产任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
        traceId:
          type: string
...
    ProductionTask:
      type: object
      properties:
        id: { type: integer, format: int64 }
        taskNo: { type: string }
        jobTicketId: { type: integer, format: int64 }
        station: { type: string }
        operatorName: { type: string }
        plannedStart: { type: string }
        plannedEnd: { type: string }
        status: { type: string }
        progressPercent: { type: integer }
        qualityStatus: { type: string }
...